+++ config/tenancy.php
@@ -9,8 +9,19 @@ return [
      * It is recommended to use the ::class property to allow auto-updating the class
      * names when using IDE refactoring tools.
      */
-    'participant_class'      => '',
-    'participant_repository' => '',
+    'participant_class'             => '',
+    'participant_repository'        => '',
+
+    /*
+     * For multi-site tenancy i.e. by domain name, specify the
+     * DomainAwareTenantParticipantRepository here. This will be used to find tenants
+     * by the domain name. The entity class is also required.
+     *
+     * Note: that the interface extends TenantParticipant so the same details can be
+     * used here as in the previous participant settings.
+     */
+    'domain_participant_class'      => null,
+    'domain_participant_repository' => null,
 
     /*
      * Map aliases to the various participant classes
@@ -31,6 +42,20 @@ return [
      * move your entity classes, these will all auto-update (if using an IDE).
      */
     'participant_mappings' => [],
+
+    /*
+     * An array of domain chunks that should be removed from the host name before
+     * performing a domain aware tenant participant look-up.
+     *
+     * For example: your main domain is www.example.com but in testing and dev it
+     * is prefixed with dev. and test. this array can contain these (and www.) and
+     * the host will always be resolved to example.com.
+     *
+     * Leave this empty to have the domain look up always check the full domain.
+     */
+    'ignorable_domain_components' => [
+        'dev.', 'test.',
+    ],
 
     /*
      * Add any repositories that need to be added to the container with the